requirements.txt
fastapi
uvicorn
python-multipart
pylatexenc
pdf2image
mistralai
langchain-google-genai
google-generativeai
langchain-community
faiss-cpu
python-dotenv
streamlit
requests